# docs/swagger.yaml
basePath: /
definitions:
...
  response.Envelope-users_UserListResponse:
    properties:
      data:
        $ref: '#/definitions/users.UserListResponse'
    type: object
  response.Envelope-users_UserResponse:
    properties:
      data:
        $ref: '#/definitions/users.UserResponse'
    type: object
  response.ErrorBody:
    properties:
      error:
        $ref: '#/definitions/response.ErrorDetail'
    type: object
  response.ErrorDetail:
    properties:
      code:
        type: string
      kind:
        type: string
      message:
        type: string
    type: object
...
  users.CreateUserRequest:
    properties:
      email:
        example: alex@acme.com
        maxLength: 320
        type: string
      first_name:
        example: Alex
        maxLength: 120
        minLength: 1
        type: string
      last_name:
        example: Rivera
        maxLength: 120
        type: string
      password:
        example: SecurePass123
        maxLength: 72
        minLength: 8
        type: string
      phone:
        example: +1-555-0199
        maxLength: 64
        type: string
      role:
        enum:
        - admin
        - landlord
        - manager
        - staff
        example: staff
        type: string
      status:
        enum:
        - active
        - disabled
        - invited
        example: active
        type: string
    required:
    - email
    - first_name
    - role
    type: object
  users.PatchUserRequest:
    properties:
      email:
        example: alex.new@acme.com
        maxLength: 320
        type: string
      first_name:
        example: Alex
        maxLength: 120
        minLength: 1
        type: string
      last_name:
        example: Rivera
        maxLength: 120
        type: string
      password:
        example: NewSecurePass456
        maxLength: 72
        minLength: 8
        type: string
      phone:
        example: +1-555-0199
        maxLength: 64
        type: string
      role:
        enum:
        - admin
        - landlord
        - manager
        - staff
        example: manager
        type: string
      status:
        enum:
        - active
        - disabled
        - invited
        example: disabled
        type: string
    type: object
  users.UserListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/users.UserResponse'
        type: array
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      total:
        example: 42
        type: integer
    type: object
  users.UserResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      email:
        example: alex@acme.com
        type: string
      first_name:
        example: Alex
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      last_name:
        example: Rivera
        type: string
      organization_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      phone:
        example: +1-555-0199
        type: string
      role:
        example: staff
        type: string
      status:
        example: active
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Rent Management System HTTP API (bootstrap; module routes are placeholders).
  title: RMS API
  version: "1.0"
paths:
...
  /api/v1/organizations/{id}/users:
    get:
      description: '**Admin:** any organization `id`. **Landlord / manager:** only
        users in their own organization (`id` must match the token). **Staff** is
        not allowed. Paginated with `page` and `page_size` (max 100).'
      parameters:
      - description: Organization ID
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page number (1-based)
        example: 1
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Page size
        example: 20
        in: query
        maximum: 100
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope-users_UserListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: List organization users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: '**Admin:** may create users in any organization with any role.
        **Landlord:** may create `manager` or `staff` in their org. **Manager:** may
        create `staff` only. Password is required unless `status` is `invited` (server
        generates a temporary password). Email must be unique per organization.'
      parameters:
      - description: Organization ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Create payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Envelope-users_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create organization user
      tags:
      - users
  /api/v1/organizations/{id}/users/{userId}:
    patch:
      consumes:
      - application/json
      description: '**Admin:** may update any user in any org. **Landlord:** may update
        `manager` and `staff` users in their org (not other landlords or admins).
        **Manager:** may update `staff` only. Optional password is re-hashed when
        provided.'
      parameters:
      - description: Organization ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        format: uuid
        in: path
        name: userId
        required: true
        type: string
      - description: 'Example: {\'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope-users_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update organization user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"